+++ app/content/signal-lessons.ts
@@ -1,0 +1,9 @@
+export const signalLessonSlugs = [
+  "digital-signals",
+  "convolution",
+  "fir-iir-filters",
+  "sinusoid-anatomy",
+  "transfer-functions",
+  "sampling-aliasing",
+  "fourier-dft",
+] as const;